api: major fix to Get(), Min(), Max() APIs.
Instead of return a Node object these APIs are not accepting
a node-callback. The node object is valid only till callback
returns and this can help in reducing garbage.

These APIs now return a boolean value, where if,
true  - key is present and returns the callback's return value
true  - key is present and callback is nil
false - key is missing

+++ llrb/metadata.go
@@ -294,8 +294,8 @@ func (md *metadata) vbuuid() uint64 {
 	return md.fields[off-1]
 }
 
-func (md *metadata) fpos() (level byte, offset uint64) {
+func (md *metadata) fpos() (level byte, offset int64) {
 	off := mdlookup[mdOffsetmaskFpos&md.hdr]
 	fpos := md.fields[off-1]
-	return byte(fpos >> 59), fpos & 0x07FFFFFFFFFFFFFF
+	return byte(fpos >> 59), int64(fpos & 0x07FFFFFFFFFFFFFF)
 }
